initial re-architecture with aie_base
