+++ src/runtime_src/xdp/profile/plugin/aie_base/generations/aie_generations.h
@@ -1,0 +1,38 @@
+// #######################################################################
+// Copyright (c) 2025 AMD, Inc.  All rights reserved.
+//
+// This   document  contains  proprietary information  which   is
+// protected by  copyright. All rights  are reserved. No  part of
+// this  document may be photocopied, reproduced or translated to
+// another  program  language  without  prior written  consent of
+// XILINX Inc., San Jose, CA. 95124
+//
+// Xilinx, Inc.
+// XILINX IS PROVIDING THIS DESIGN, CODE, OR INFORMATION "AS IS" AS A
+// COURTESY TO YOU.  BY PROVIDING THIS DESIGN, CODE, OR INFORMATION AS
+/// ONE POSSIBLE   IMPLEMENTATION OF THIS FEATURE, APPLICATION OR
+// STANDARD, XILINX IS MAKING NO REPRESENTATION THAT THIS IMPLEMENTATION
+// IS FREE FROM ANY CLAIMS OF INFRINGEMENT, AND YOU ARE RESPONSIBLE
+// FOR OBTAINING ANY RIGHTS YOU MAY REQUIRE FOR YOUR IMPLEMENTATION.
+// XILINX EXPRESSLY DISCLAIMS ANY WARRANTY WHATSOEVER WITH RESPECT TO
+// THE ADEQUACY OF THE IMPLEMENTATION, INCLUDING BUT NOT LIMITED TO
+// ANY WARRANTIES OR REPRESENTATIONS THAT THIS IMPLEMENTATION IS FREE
+// FROM CLAIMS OF INFRINGEMENT, IMPLIED WARRANTIES OF MERCHANTABILITY
+// AND FITNESS FOR A PARTICULAR PURPOSE.
+//
+// ######################################################################
+
+#ifndef AIE_GENERATIONS_H_
+#define AIE_GENERATIONS_H_
+
+#include "aie1_attributes.h"
+#include "aie2_attributes.h"
+#include "aie2ps_attributes.h"
+#include "npu3_attributes.h"
+
+#include "aie1_registers.h"
+#include "aie2_registers.h"
+#include "aie2ps_registers.h"
+#include "npu3_registers.h"
+
+#endif /* AIE_GENERATIONS_H_ */
