+++ src/runtime_src/xdp/profile/plugin/aie_debug/CMakeLists.txt
@@ -17,9 +17,9 @@ elseif (DEFINED XRT_AIE_BUILD AND ${XRT_NATIVE_BUILD} STREQUAL "no")
 endif()
 
 file(GLOB AIE_DEBUG_PLUGIN_FILES
+  "${PROFILE_DIR}/plugin/aie_base/generations/*.h"
   "${PROFILE_DIR}/plugin/aie_debug/*.h"
   "${PROFILE_DIR}/plugin/aie_debug/*.cpp"
-  "${PROFILE_DIR}/plugin/aie_debug/generations/*.h"
   "${PROFILE_DIR}/writer/aie_debug/*.h"
   "${PROFILE_DIR}/writer/aie_debug/*.cpp"
   "${IMPL_DIR}/*.h"
