+++ src/runtime_src/xdp/profile/plugin/aie_profile/util/aie_profile_util.h
@@ -113,13 +113,6 @@ namespace xdp::aie::profile {
                     std::vector<XAie_Events>& events, const int hwGen);
 
   /**
-   * @brief  Check if event is generated by a stream switch monitor port
-   * @param  event Event ID to check
-   * @return True if given event is from a stream switch port
-   */
-  bool isStreamSwitchPortEvent(const XAie_Events event);
-
-  /**
    * @brief Configure the individual AIE events for metric sets that use group events
    * @param aieDevInst AIE device instance
    * @param loc Tile location
@@ -145,36 +138,6 @@ namespace xdp::aie::profile {
   void configEventSelections(XAie_DevInst* aieDevInst, const XAie_LocType loc,
                              const module_type type, const std::string metricSet,
                              const uint8_t channel);
-
-  /**
-   * @brief  Check if event is a port running event
-   * @param  event Event ID to check
-   * @return True if given event is a port running event
-   */
-  bool isPortRunningEvent(const XAie_Events event);
-
-  /**
-   * @brief  Check if event is a port stalled event
-   * @param  event Event ID to check
-   * @return True if given event is a port stalled event
-   */
-  bool isPortStalledEvent(const XAie_Events event);
-
-  /**
-   * @brief  Check if event is a port idle event
-   * @param  event Event ID to check
-   * @return True if given event is a port idle event
-   */
-  bool isPortIdleEvent(const XAie_Events event);
-
-  /**
-   * @brief  Check if event is a port tlast event
-   * @param  event Event ID to check
-   * @return True if given event is a port tlast event
-   */
-  bool isPortTlastEvent(const XAie_Events event);
-
-  uint8_t getPortNumberFromEvent(const XAie_Events event);
   
   /**
    * @brief Get XAie module enum at the module index
