+++ src/runtime_src/xdp/profile/plugin/aie_trace/client/aie_trace.h
@@ -53,13 +53,7 @@ namespace xdp {
       uint32_t bcIdToEvent(int bcId);
       
       bool isInputSet(const module_type type, const std::string metricSet);
-      bool isStreamSwitchPortEvent(const XAie_Events event);
-      bool isPortRunningEvent(const XAie_Events event);
-      bool isCoreModuleEvent(const XAie_Events event);
-      bool isDmaSet(const std::string metricSet);
 
-      uint8_t getPortNumberFromEvent(XAie_Events event);
-      int8_t getChannelNumberFromEvent(XAie_Events event);
       void configStreamSwitchPorts(const tile_type& tile, const XAie_LocType loc,
                                    const module_type type, const std::string metricSet, 
                                    const uint8_t channel0, const uint8_t channel1,
