+++ src/runtime_src/xdp/profile/plugin/aie_trace/util/aie_trace_config.cpp
@@ -100,7 +100,7 @@ namespace xdp::aie::trace {
     for (int i=0; i < events.size(); ++i) {
       // Ensure applicable event
       auto event = events.at(i);
-      if (!isStreamSwitchPortEvent(event))
+      if (!xdp::aie::isStreamSwitchPortEvent(event))
         continue;
 
       bool newPort = false;
@@ -179,7 +179,7 @@ namespace xdp::aie::trace {
       // Event options:
       //   getSSIdleEvent, getSSRunningEvent, getSSStalledEvent, & getSSTlastEvent
       XAie_Events ssEvent;
-      if (isPortRunningEvent(event))
+      if (aie::isPortRunningEvent(event))
         switchPortRsc->getSSRunningEvent(ssEvent);
       else
         switchPortRsc->getSSStalledEvent(ssEvent);
@@ -213,7 +213,7 @@ namespace xdp::aie::trace {
                     aie_cfg_base& config)
   {
     // Only needed for core/memory modules and metric sets that include DMA events
-    if (!isDmaSet(metricSet) || ((type != module_type::core) && (type != module_type::dma)))
+    if (!xdp::aie::isDmaSet(metricSet) || ((type != module_type::core) && (type != module_type::dma)))
       return {};
 
     std::vector<XAie_Events> comboEvents;
@@ -288,7 +288,7 @@ namespace xdp::aie::trace {
                          const std::string metricSet)
   {
     // Only needed for core module and metric sets that include DMA events
-    if (!isDmaSet(metricSet) || (type != module_type::core))
+    if (!xdp::aie::isDmaSet(metricSet) || (type != module_type::core))
       return;
 
     // Set masks for group events
