+++ src/runtime_src/xdp/profile/plugin/aie_trace/util/aie_trace_util.cpp
@@ -311,19 +311,6 @@ namespace xdp::aie::trace {
   }
 
   /****************************************************************************
-   * Check if metric set contains DMA events
-   * TODO: Traverse events vector instead of based on name
-   ***************************************************************************/
-  bool isDmaSet(const std::string metricSet)
-  {
-    if ((metricSet.find("dma") != std::string::npos)
-        || (metricSet.find("s2mm") != std::string::npos)
-        || (metricSet.find("mm2s") != std::string::npos))
-      return true;
-    return false;
-  }
-
-  /****************************************************************************
    * Get start events for core module counters
    ***************************************************************************/
   std::vector<XAie_Events> getCoreCounterStartEvents(int hwGen, std::string scheme)
@@ -425,185 +412,6 @@ namespace xdp::aie::trace {
     else if (scheme == "es2")
       eventValues = {ES2_TRACE_COUNTER};
     return eventValues;
-  }
-
-  /****************************************************************************
-   * Check if core module event
-   ***************************************************************************/
-  bool isCoreModuleEvent(const XAie_Events event)
-  {
-    return ((event >= XAIE_EVENT_NONE_CORE) 
-            && (event <= XAIE_EVENT_INSTR_ERROR_CORE));
-  }
-
-  /****************************************************************************
-   * Check if stream switch port event
-   ***************************************************************************/
-  bool isStreamSwitchPortEvent(const XAie_Events event)
-  {
-    // AIE tiles
-    if ((event > XAIE_EVENT_GROUP_STREAM_SWITCH_CORE) 
-        && (event < XAIE_EVENT_GROUP_BROADCAST_CORE))
-      return true;
-    // Interface tiles
-    if ((event > XAIE_EVENT_GROUP_STREAM_SWITCH_PL) 
-        && (event < XAIE_EVENT_GROUP_BROADCAST_A_PL))
-      return true;
-    // Memory tiles
-    if ((event > XAIE_EVENT_GROUP_STREAM_SWITCH_MEM_TILE) 
-        && (event < XAIE_EVENT_GROUP_MEMORY_CONFLICT_MEM_TILE))
-      return true;
-
-    return false;
-  }
-
-  /****************************************************************************
-   * Check if port running event
-   ***************************************************************************/
-  bool isPortRunningEvent(const XAie_Events event)
-  {
-    std::set<XAie_Events> runningEvents = {
-      XAIE_EVENT_PORT_RUNNING_0_CORE,     XAIE_EVENT_PORT_RUNNING_1_CORE,
-      XAIE_EVENT_PORT_RUNNING_2_CORE,     XAIE_EVENT_PORT_RUNNING_3_CORE,
-      XAIE_EVENT_PORT_RUNNING_4_CORE,     XAIE_EVENT_PORT_RUNNING_5_CORE,
-      XAIE_EVENT_PORT_RUNNING_6_CORE,     XAIE_EVENT_PORT_RUNNING_7_CORE,
-      XAIE_EVENT_PORT_RUNNING_0_PL,       XAIE_EVENT_PORT_RUNNING_1_PL,
-      XAIE_EVENT_PORT_RUNNING_2_PL,       XAIE_EVENT_PORT_RUNNING_3_PL,
-      XAIE_EVENT_PORT_RUNNING_4_PL,       XAIE_EVENT_PORT_RUNNING_5_PL,
-      XAIE_EVENT_PORT_RUNNING_6_PL,       XAIE_EVENT_PORT_RUNNING_7_PL,
-      XAIE_EVENT_PORT_RUNNING_0_MEM_TILE, XAIE_EVENT_PORT_RUNNING_1_MEM_TILE,
-      XAIE_EVENT_PORT_RUNNING_2_MEM_TILE, XAIE_EVENT_PORT_RUNNING_3_MEM_TILE,
-      XAIE_EVENT_PORT_RUNNING_4_MEM_TILE, XAIE_EVENT_PORT_RUNNING_5_MEM_TILE,
-      XAIE_EVENT_PORT_RUNNING_6_MEM_TILE, XAIE_EVENT_PORT_RUNNING_7_MEM_TILE
-    };
-
-    return (runningEvents.find(event) != runningEvents.end());
-  }
-
-  /****************************************************************************
-   * Get port number based on event
-   ***************************************************************************/
-  uint8_t getPortNumberFromEvent(XAie_Events event)
-  {
-    switch (event) {
-    case XAIE_EVENT_PORT_RUNNING_3_CORE:
-    case XAIE_EVENT_PORT_STALLED_3_CORE:
-    case XAIE_EVENT_PORT_IDLE_3_CORE:
-    case XAIE_EVENT_PORT_RUNNING_3_PL:
-    case XAIE_EVENT_PORT_STALLED_3_PL:
-    case XAIE_EVENT_PORT_IDLE_3_PL:
-      return 3;
-    case XAIE_EVENT_PORT_RUNNING_2_CORE:
-    case XAIE_EVENT_PORT_STALLED_2_CORE:
-    case XAIE_EVENT_PORT_IDLE_2_CORE:
-    case XAIE_EVENT_PORT_RUNNING_2_PL:
-    case XAIE_EVENT_PORT_STALLED_2_PL:
-    case XAIE_EVENT_PORT_IDLE_2_PL:
-      return 2;
-    case XAIE_EVENT_PORT_RUNNING_1_CORE:
-    case XAIE_EVENT_PORT_STALLED_1_CORE:
-    case XAIE_EVENT_PORT_IDLE_1_CORE:
-    case XAIE_EVENT_PORT_RUNNING_1_PL:
-    case XAIE_EVENT_PORT_STALLED_1_PL:
-    case XAIE_EVENT_PORT_IDLE_1_PL:
-      return 1;
-    default:
-      return 0;
-    }
-  }
-
-  /****************************************************************************
-   * Get channel number based on event
-   * NOTE: This only covers AIE Tiles and Interface Tiles
-   ***************************************************************************/
-  int8_t getChannelNumberFromEvent(XAie_Events event)
-  {
-    switch (event) {
-    case XAIE_EVENT_DMA_S2MM_0_START_TASK_MEM:
-    case XAIE_EVENT_DMA_S2MM_0_FINISHED_BD_MEM:
-    case XAIE_EVENT_DMA_S2MM_0_FINISHED_TASK_MEM:
-    case XAIE_EVENT_DMA_S2MM_0_STALLED_LOCK_MEM:
-    case XAIE_EVENT_DMA_S2MM_0_STREAM_STARVATION_MEM:
-    case XAIE_EVENT_DMA_S2MM_0_MEMORY_BACKPRESSURE_MEM:
-    case XAIE_EVENT_DMA_MM2S_0_START_TASK_MEM:
-    case XAIE_EVENT_DMA_MM2S_0_FINISHED_BD_MEM:
-    case XAIE_EVENT_DMA_MM2S_0_FINISHED_TASK_MEM:
-    case XAIE_EVENT_DMA_MM2S_0_STALLED_LOCK_MEM:
-    case XAIE_EVENT_DMA_MM2S_0_STREAM_BACKPRESSURE_MEM:
-    case XAIE_EVENT_DMA_MM2S_0_MEMORY_STARVATION_MEM:
-    case XAIE_EVENT_DMA_S2MM_0_START_BD_PL:
-    case XAIE_EVENT_DMA_S2MM_0_FINISHED_BD_PL:
-    case XAIE_EVENT_DMA_S2MM_0_START_TASK_PL:
-    case XAIE_EVENT_DMA_S2MM_0_FINISHED_TASK_PL:
-    case XAIE_EVENT_DMA_S2MM_0_STALLED_LOCK_PL:
-    case XAIE_EVENT_DMA_S2MM_0_STREAM_STARVATION_PL:
-    case XAIE_EVENT_DMA_S2MM_0_MEMORY_BACKPRESSURE_PL:
-    case XAIE_EVENT_DMA_MM2S_0_START_BD_PL:
-    case XAIE_EVENT_DMA_MM2S_0_FINISHED_BD_PL:
-    case XAIE_EVENT_DMA_MM2S_0_START_TASK_PL:
-    case XAIE_EVENT_DMA_MM2S_0_FINISHED_TASK_PL:
-    case XAIE_EVENT_DMA_MM2S_0_STALLED_LOCK_PL:
-    case XAIE_EVENT_DMA_MM2S_0_STREAM_BACKPRESSURE_PL:
-    case XAIE_EVENT_DMA_MM2S_0_MEMORY_STARVATION_PL:
-#ifdef XDP_VE2_BUILD
-    case XAIE_EVENT_NOC0_DMA_S2MM_0_START_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_0_FINISHED_BD_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_0_FINISHED_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_0_STALLED_LOCK_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_0_STREAM_STARVATION_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_0_MEMORY_BACKPRESSURE_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_0_START_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_0_FINISHED_BD_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_0_FINISHED_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_0_STALLED_LOCK_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_0_STREAM_BACKPRESSURE_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_0_MEMORY_STARVATION_PL:
-#endif
-      return 0;
-    case XAIE_EVENT_DMA_S2MM_1_START_TASK_MEM:
-    case XAIE_EVENT_DMA_S2MM_1_FINISHED_BD_MEM:
-    case XAIE_EVENT_DMA_S2MM_1_FINISHED_TASK_MEM:
-    case XAIE_EVENT_DMA_S2MM_1_STALLED_LOCK_MEM:
-    case XAIE_EVENT_DMA_S2MM_1_STREAM_STARVATION_MEM:
-    case XAIE_EVENT_DMA_S2MM_1_MEMORY_BACKPRESSURE_MEM:
-    case XAIE_EVENT_DMA_MM2S_1_START_TASK_MEM:
-    case XAIE_EVENT_DMA_MM2S_1_FINISHED_BD_MEM:
-    case XAIE_EVENT_DMA_MM2S_1_FINISHED_TASK_MEM:
-    case XAIE_EVENT_DMA_MM2S_1_STALLED_LOCK_MEM:
-    case XAIE_EVENT_DMA_MM2S_1_STREAM_BACKPRESSURE_MEM:
-    case XAIE_EVENT_DMA_MM2S_1_MEMORY_STARVATION_MEM:
-    case XAIE_EVENT_DMA_S2MM_1_START_BD_PL:
-    case XAIE_EVENT_DMA_S2MM_1_FINISHED_BD_PL:
-    case XAIE_EVENT_DMA_S2MM_1_START_TASK_PL:
-    case XAIE_EVENT_DMA_S2MM_1_FINISHED_TASK_PL:
-    case XAIE_EVENT_DMA_S2MM_1_STALLED_LOCK_PL:
-    case XAIE_EVENT_DMA_S2MM_1_STREAM_STARVATION_PL:
-    case XAIE_EVENT_DMA_S2MM_1_MEMORY_BACKPRESSURE_PL:
-    case XAIE_EVENT_DMA_MM2S_1_START_BD_PL:
-    case XAIE_EVENT_DMA_MM2S_1_FINISHED_BD_PL:
-    case XAIE_EVENT_DMA_MM2S_1_START_TASK_PL:
-    case XAIE_EVENT_DMA_MM2S_1_FINISHED_TASK_PL:
-    case XAIE_EVENT_DMA_MM2S_1_STALLED_LOCK_PL:
-    case XAIE_EVENT_DMA_MM2S_1_STREAM_BACKPRESSURE_PL:
-    case XAIE_EVENT_DMA_MM2S_1_MEMORY_STARVATION_PL:
-#ifdef XDP_VE2_BUILD
-    case XAIE_EVENT_NOC0_DMA_S2MM_1_START_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_1_FINISHED_BD_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_1_FINISHED_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_1_STALLED_LOCK_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_1_STREAM_STARVATION_PL:
-    case XAIE_EVENT_NOC0_DMA_S2MM_1_MEMORY_BACKPRESSURE_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_1_START_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_1_FINISHED_BD_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_1_FINISHED_TASK_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_1_STALLED_LOCK_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_1_STREAM_BACKPRESSURE_PL:
-    case XAIE_EVENT_NOC0_DMA_MM2S_1_MEMORY_STARVATION_PL:
-#endif
-      return 1;
-    default:
-      return -1;
-    }
   }
 
   /****************************************************************************
