+++ src/runtime_src/xdp/profile/plugin/aie_trace/util/aie_trace_util.h
@@ -117,48 +117,6 @@ namespace xdp::aie::trace {
   std::vector<uint32_t> getMemoryCounterEventValues(int hwGen, std::string scheme);
 
   /**
-   * @brief  Check if metric set contains DMA events
-   * @param  metricSet Name of requested metric set
-   * @return True if given metric set contains DMA event(s)
-   */
-  bool isDmaSet(const std::string metricSet);
-
-  /**
-   * @brief  Check if event is core module event
-   * @param  event Event ID to check
-   * @return True if given event is from a core module
-   */
-  bool isCoreModuleEvent(const XAie_Events event);
-
-  /**
-   * @brief  Check if event is generated by a stream switch monitor port
-   * @param  event Event ID to check
-   * @return True if given event is from a stream switch port
-   */
-  bool isStreamSwitchPortEvent(const XAie_Events event);
-
-  /**
-   * @brief  Check if event is a port running event
-   * @param  event Event ID to check
-   * @return True if given event is a port running event
-   */
-  bool isPortRunningEvent(const XAie_Events event);
-
-  /**
-   * @brief  Get port number from event
-   * @param  event Event ID to check
-   * @return Port number associated with given event (default: 0)
-   */
-  uint8_t getPortNumberFromEvent(XAie_Events event);
-
-  /**
-   * @brief  Get channel number from event
-   * @param  event Event ID to check
-   * @return Channel number associated with given event (default: -1)
-   */
-  int8_t getChannelNumberFromEvent(XAie_Events event);
-
-  /**
    * @brief Print out trace event statistics across multiple tiles
    * @param m        Module index (used to get name)
    * @param numTiles Array of tile usage stats (length: NUM_TRACE_EVENTS)
